Add listfolder_api my sample code and updown dropbox's sample code
updown.py is dropbox's sample code that pulls and pushes a download folder

I am trying stuff out in listfolder_api to understand
how to do things.  The first step is to be able to get
the total size of dropbox and the number of files
correctly.  Then, I need to identify which files were
made at what date.  After, I need to pull, push
a file, and delete files.

I should make this into some sort of tree.  Go through
and do this for each branch of the tree.  Easier to organize
and eliminates the weird problem.